pyproject.toml
# ABOUTME: Project configuration for feed-brain, AI-powered personal feed aggregator.
# ABOUTME: Defines dependencies, build system, and tool configurations.

[project]
name = "feed-brain"
version = "0.1.0"
description = "AI-powered personal feed aggregator with LLM classification"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [{ name = "Max" }]

dependencies = [
    "anthropic>=0.40.0",
    "feedparser>=6.0.11",
    "httpx>=0.27.0",
    "readability-lxml>=0.8.1",
    "beautifulsoup4>=4.12.3",
    "lxml-html-clean>=0.4.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "structlog>=25.0.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    # Web
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "Jinja2>=3.1.5",
    "python-multipart>=0.0.22",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.9.0",
    "respx>=0.22.0",
]

[project.scripts]
feed-brain = "feed_brain.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/feed_brain"]

[tool.ruff]
line-length = 100
target-version = "py313"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["feed_brain"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["src/feed_brain"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
